refactor(strategy): extract indicators to strategy/indicators.py (#186 A2)
