+++ strategy/__init__.py
@@ -1,0 +1,1 @@
+"""Strategy module — pure trading logic shared by scanner and backtest (Epic #186)."""
